changed structure, added layouts
