+++ templates/login.html
@@ -1,12 +1,4 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css">
-    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
-    <title>Login Page</title>
-</head>
+{% extends "login_layout.html" %}
 <body>
     <a href="/" style="color: black;">← Go Back</a>
     <div class="login-container" style="height: 100%;">
@@ -14,6 +6,7 @@
             <div class="logo">
                 <img src="{{ url_for('static', filename='cit-logo-white.png') }}" alt="Bootstrap" width="100" height="80">
             </div>
+{% block content %}
             <h3>NIM</h3>
             <form method="POST" action="/login_user">
             <div class="nim-input">
@@ -34,9 +27,5 @@
             </form>
         </div>
     </div>
-    
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
-    <script src="https://kit.fontawesome.com/48638eddb5.js" crossorigin="anonymous"></script>
-    <script src="{{ url_for('static', filename='script.js') }}"></script>
-</body>
-</html>
+
+{% endblock content %}
